+++ src/main/java/com/example/demo/repositories/LeaseRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.services;
+package com.example.demo.repositories;
 
 import com.example.demo.models.LeaseModel;
 import org.springframework.data.mongodb.repository.MongoRepository;
