+++ src/main/java/com/example/demo/services/Leasing.java
@@ -1,6 +1,7 @@
 package com.example.demo.services;
 
 import com.example.demo.models.LeaseModel;
+import com.example.demo.repositories.LeaseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
